+++ session/src/init/templates/config/kv.ts
@@ -18,6 +18,12 @@ export const sessionConfig: SessionOptions = {
     sameSite: "Lax",
     maxAge: 60 * 60 * 24 * 7, // 1 week
   },
+  // Enable for Stateless API Token Support (e.g. "Authorization: Bearer <token>")
+  // verifyToken: async (token) => {
+  //   // const user = await findUserByToken(token);
+  //   // return user;
+  // },
+  // tokenPrefix: "Bearer ", // Optional (Default: "Bearer ")
 };
 
 export const session = createSessionMiddleware<State>(sessionConfig);