+++ session/src/init/templates/config/kvdex.ts
@@ -1,9 +1,13 @@
-import { createSessionMiddleware } from "@innovatedev/fresh-session";
+import {
+  createSessionMiddleware,
+  type SessionOptions,
+} from "@innovatedev/fresh-session";
 import { collection, kvdex, type KvValue, model } from "@olli/kvdex";
 import {
   KvDexSessionStorage,
   type SessionDoc,
 } from "@innovatedev/fresh-session/kvdex-store";
+import type { State } from "../utils.ts";
 
 const kv = await Deno.openKv();
 
@@ -28,15 +32,26 @@ const db = kvdex({
   },
 });
 
-export const session = createSessionMiddleware({
+export const sessionConfig: SessionOptions = {
   store: new KvDexSessionStorage({
     collection: db.sessions,
     userCollection: db.users,
+    expireAfter: 60 * 60 * 24 * 7, // 1 week
+    // userIndex: "email", // Optional secondary index
   }),
+  cookie: {
+    name: "sessionId",
+    httpOnly: true,
+    secure: true,
+    sameSite: "Lax",
+    maxAge: 60 * 60 * 24 * 7, // 1 week
+  },
   // Enable for Stateless API Token Support (e.g. "Authorization: Bearer <token>")
   // verifyToken: async (token) => {
   //   // const user = await findUserByToken(token);
   //   // return user;
   // },
   // tokenPrefix: "Bearer ", // Optional (Default: "Bearer ")
-});
+};
+
+export const session = createSessionMiddleware<State>(sessionConfig);